+++ Cargo.toml
@@ -2,5 +2,4 @@
 members = [
   "src/token",
   "src/factory",
-  "src/common",
 ]
