--- src/common/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-edition = "2021"
-name = "common"
-version = "0.1.2"
-
-[dependencies]
-candid = "0.7"
-ic-cdk = "0.5"
-serde = "1.0"
-#ic-helpers = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-helpers", tag = "v0.3.4" }
-ic-helpers = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-helpers", branch = "maxim/CPROD-801-tokens_wrappers" }
-#ic-helpers = {path = "../../../ic-helpers/ic-helpers"}
