+++ src/token/Cargo.toml
@@ -10,7 +10,6 @@ no_api = []
 [dependencies]
 assert-panic = "1.0"
 candid = "0.7"
-common = {path = "../common"}
 ic-cdk = "0.5"
 ic-cdk-macros = "0.5"
 num-traits = "0.2"
@@ -18,12 +17,12 @@ serde = "1.0"
 #ic-canister = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-canister", tag = "v0.3.4" }
 #ic-helpers = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-helpers", tag = "v0.3.4" }
 #ic-storage = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-storage", tag = "v0.3.4" }
-ic-canister = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-canister", branch = "maxim/CPROD-801-tokens_wrappers" }
-ic-helpers = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-helpers", branch = "maxim/CPROD-801-tokens_wrappers" }
-ic-storage = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-storage", branch = "maxim/CPROD-801-tokens_wrappers" }
-#ic-canister = {path = "../../../ic-helpers/ic-canister/ic-canister"}
-#ic-helpers = {path = "../../../ic-helpers/ic-helpers"}
-#ic-storage = {path = "../../../ic-helpers/ic-storage"}
+#ic-canister = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-canister", branch = "maxim/CPROD-801-tokens_wrappers" }
+#ic-helpers = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-helpers", branch = "maxim/CPROD-801-tokens_wrappers" }
+#ic-storage = { git = "https://github.com/infinity-swap/ic-helpers", package = "ic-storage", branch = "maxim/CPROD-801-tokens_wrappers" }
+ic-canister = {path = "../../../ic-helpers/ic-canister/ic-canister"}
+ic-helpers = {path = "../../../ic-helpers/ic-helpers"}
+ic-storage = {path = "../../../ic-helpers/ic-storage"}
 ic-certified-map = "0.3"
 serde_cbor = "0.11"
